ath9k: Add support for reading the EEPROM data using the nvmem API
Currently SoC platforms use a firmware request to get the EEPROM data.
This is mostly a hack and rely on using a user-helper scripts which is
deprecated. A nicer alternative is to use the nvmem API which was
designed for this kind of task.

Furthermore we let CONFIG_ATH9K_AHB select CONFIG_NVMEM as such
devices will generally use this method for loading the EEPROM data.

Signed-off-by: Alban Bedel <albeu@free.fr>
---
Changelog:
v2: * Fixed From/SoB to have my full name
    * Added a mssing new line in the eeprom read error message